删除SofeReference,完成缓存的持久化和磁盘读取

+++ joice-cache/src/main/java/org/joice/cache/Cache.java
@@ -31,8 +31,8 @@ public interface Cache {
     /**
      * 删除缓存
      * @param cacheKey  缓存的key,不能为空 
-     * @return          缓存变化的数量
+     * @return          删除缓存的数量
      */
-    int delete(CacheKey cacheKey);
+    Long delete(CacheKey cacheKey);
 
 }
